zoomtoics/requirements.txt
arrow==0.14.7
boto3==1.17.60
botocore==1.20.60
certifi==2020.12.5
chardet==4.0.0
ics==0.7
idna==2.10
jmespath==0.10.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
TatSu==5.6.1
urllib3==1.26.3
zoomus==1.1.4
